backend/requirements.txt
# Star Auto - Django Backend

## Requirements
Django>=4.2,<5.0
djangorestframework>=3.14,<4.0
django-cors-headers>=4.3,<5.0
djangorestframework-simplejwt>=5.3,<6.0
Pillow>=10.0,<11.0
python-dotenv>=1.0,<2.0
gunicorn>=21.0,<22.0

# Database
# Use SQLite for development, PostgreSQL for production
# psycopg2-binary>=2.9,<3.0  # Uncomment for PostgreSQL
